requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
isort==4.3.21
Pillow==7.0.0
python-dotenv==0.12.0
pytz==2019.3
sorl-thumbnail==12.6.3
sqlparse==0.3.0
